refactor: move node configuration from gemstone to primitives module
