+++ Cargo.toml
@@ -129,6 +129,12 @@ uniffi_bindgen = { version = "0.29.2" }
 alloy-primitives = "1.1.0"
 alloy-sol-types = { version = "1.1.0", features = ["eip712-serde"] }
 alloy-dyn-abi = { version = "1.1.0", features = ["eip712"] }
+alloy-ens = "1.0.9"
+alloy-json-rpc = "1.0.9"
+alloy-rpc-client = "1.0.9"
+alloy-rpc-types = "1.0.9"
+alloy-transport-http = "1.0.9"
+alloy-contract = "1.0.9"
 
 # localizer
 i18n-embed = { version = "0.15.4", features = ["fluent-system", "autoreload"] }
